requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.40.0
requests>=2.31.0
tiktoken>=0.5.0
